cleaning up tests directory
