+++ test/components/form_test.js
@@ -1,0 +1,9 @@
+var expect = require('expect');
+
+describe("Form", function(){
+  context("when visited", function(){
+    it("should contain empty input values", function(){
+      //expect(1).toEqual(2)
+    })
+  })
+})
